refactoring backbone model specs
